refactor: extract ScanFlowView from onboarding ScanStepView
